Refactor ParsePortSpec to handle IPv6 addresses
The PortSpec uses colons a delimiter between parts. If an IPv6
address is used, this leads to the colons in the IP-address to
be handled as delimiter.

Commit 1d78bed8423d3377e04c4bdd1e1a045fb42e2942 also added support
for "ChildIP" to the portspec format, which (adding to the above)
made parsing the format even more ambiguous.

This patch rewrites ParsePortSpec to handle IPv6 addresses, but
requires those addresses to be wrapped by square brackes.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>